backend/requirements.txt
speechbrain==1.0.0
transformers==4.39.3
sacrebleu==2.4.2
datasets==2.18.0
fastapi==0.104.1
uvicorn==0.24.0.post1
elevenlabs==1.2.2
python-dotenv==1.0.1
openai==1.30.1